# dashboard/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Industrial Time-Series Monitoring API",
    "description": "\nREST API over a historical archive of condition-monitoring data from a metro train's\ncompressor Air Production Unit (UCI MetroPT-3, DOI 10.24432/C5VW3R).\n\n**What the data is.** 22,754,220 readings - 1,516,948 scans of 15 sensors, February to\nSeptember 2020. Coverage is 82.4%: the remaining 17.6% of the timeline has no data, in\n331 recorded gaps.\n\n**Three things this API will not do to you.**\n\n* *Silently truncate.* Any series that hits a result cap is returned with\n  `truncated: true`.\n* *Hide data quality.* 3.35% of the archive is held data from a logger freeze - non-null,\n  in range, and not a measurement. It is returned with its quality code, never filtered\n  out behind your back.\n* *Return an unusable number of points.* `resolution=auto` picks raw, hourly or daily\n  from the width of the window.\n\n**Thresholds are data, not code.** Every severity boundary traces to a row in\n`analytics.SensorBaseline` or to a setpoint published by the equipment owner.\n",
    "contact": {
      "name": "Project documentation",
      "url": "https://archive.ics.uci.edu/dataset/791/metropt+3+dataset"
    },
    "license": {
      "name": "Dataset: UCI MetroPT-3, DOI 10.24432/C5VW3R"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/equipment/{equipment_id}": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Get one piece of equipment",
        "operationId": "get_equipment_equipment__equipment_id__get",
        "parameters": [
          {
            "name": "equipment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Equipment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipment"
                }
              }
            }
          },
          "404": {
            "description": "No such resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/sensors": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "List sensors",
        "description": "Sensor definitions, including the verbatim description from the dataset docs.",
        "operationId": "list_sensors_sensors_get",
        "parameters": [
          {
            "name": "equipment_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equipment Id"
            }
          },
          {
            "name": "sensor_class",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(Analogue|Digital)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sensor Class"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sensor"
                  },
                  "title": "Response List Sensors Sensors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sensors/{sensor_id}": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Get one sensor",
        "operationId": "get_sensor_sensors__sensor_id__get",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Numeric id or sensor code, e.g. 15 or TP3.",
              "title": "Sensor Id"
            },
            "description": "Numeric id or sensor code, e.g. 15 or TP3."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sensor"
                }
              }
            }
          },
          "404": {
            "description": "No such resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Equipment": {
        "properties": {
          "equipment_id": {
            "type": "integer",
            "title": "Equipment Id"
          },
          "equipment_code": {
            "type": "string",
            "title": "Equipment Code"
          },
          "equipment_name": {
            "type": "string",
            "title": "Equipment Name"
          },
          "equipment_type": {
            "type": "string",
            "title": "Equipment Type"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "sensor_count": {
            "type": "integer",
            "title": "Sensor Count"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "equipment_id",
          "equipment_code",
          "equipment_name",
          "equipment_type",
          "sensor_count",
          "is_active"
        ],
        "title": "Equipment"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetail",
        "description": "RFC 7807 problem details.\n\nA machine-readable error shape, so a client can branch on ``type`` rather than\npattern-matching prose that will be reworded eventually.",
        "example": {
          "detail": "start (2020-06-01T00:00:00) must be earlier than end (2020-05-01T00:00:00).",
          "instance": "/readings/14",
          "status": 422,
          "title": "Invalid time range",
          "type": "/problems/invalid-time-range"
        }
      },
...
      "Sensor": {
        "properties": {
          "sensor_id": {
            "type": "integer",
            "title": "Sensor Id"
          },
          "equipment_id": {
            "type": "integer",
            "title": "Equipment Id"
          },
          "sensor_code": {
            "type": "string",
            "title": "Sensor Code"
          },
          "sensor_name": {
            "type": "string",
            "title": "Sensor Name"
          },
          "sensor_class": {
            "type": "string",
            "enum": [
              "Analogue",
              "Digital"
            ],
            "title": "Sensor Class"
          },
          "measurement_type": {
            "type": "string",
            "title": "Measurement Type"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "physical_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Physical Min"
          },
          "physical_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Physical Max"
          },
          "documented_setpoint": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documented Setpoint",
            "description": "A setpoint published by the equipment owner (e.g. the 7 bar low-pressure switch), not a threshold chosen by this project."
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Verbatim description from the dataset documentation."
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "sensor_id",
          "equipment_id",
          "sensor_code",
          "sensor_name",
          "sensor_class",
          "measurement_type",
          "description",
          "is_active"
        ],
        "title": "Sensor"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}